refactor: remove unnecessary packages and update dependencies in pubspec.yaml
This commit removes a number of redundant development dependencies from `pubspec.lock` and `pubspec.yaml`. These include `build_runner` along with many of its transitive dependencies. This change also updates the dependencies list by removing some of the packages.
The changes in `pubspec.lock` reflect a reduction in the overall number of package dependencies.
The `flutter_lints` is now set to `^5.0.0` version.